parse: finish `AyaGKParserImpl`!
